consolidated upload tests
